Changes to the hit test code so that it will work with wxPython2.8 on OS-X
This required a re-factor of the code -- the Hit Test MemoryDC is no longer kept around,
only the Bitmap is stored. To find the color hit on the hit-test bitmap, there are now
two methods -- one for 2.8, and one for earlier versions.

With 2.8, the bitmap data is accessed with wx.AlphaPixelData. On older versions, a 
wx.MemoryDC is created, and dc.GetPixelPoint is used.

So far, only tested on OS-X



git-svn-id: http://svn.wxwidgets.org/svn/wx/wxPython@51684 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775